+++ CHANGELOG.md
@@ -9,6 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Upcoming changes...
 
+## [1.20.0] - 2025-02-02
+### Added
+- Added support for component provenance reporting
+  - `scanoss-py component prov ...`
+
 ## [1.19.6] - 2025-01-30
 ### Added
 - Omit settings file if it does not exist instead of throwing an error.
@@ -451,3 +456,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [1.19.3]: https://github.com/scanoss/scanoss.py/compare/v1.19.2...v1.19.3
 [1.19.4]: https://github.com/scanoss/scanoss.py/compare/v1.19.3...v1.19.4
 [1.19.5]: https://github.com/scanoss/scanoss.py/compare/v1.19.4...v1.19.5
+[1.20.0]: https://github.com/scanoss/scanoss.py/compare/v1.19.5...v1.20.0